+++ core/src/ru/geekbrains/StarFighter.java
@@ -2,6 +2,8 @@ package ru.geekbrains;
 
 import com.badlogic.gdx.Game;
 
+import ru.geekbrains.controllers.ScreenController;
+
 
 public class StarFighter extends Game {
