+++ core/src/ru/geekbrains/base/Layer.java
@@ -2,5 +2,6 @@ package ru.geekbrains.base;
 
 public enum Layer {
     FOREGROUND,
+    MIDDLE,
     BACKGROUND;
 }
