+++ core/src/ru/geekbrains/base/Sprite.java
@@ -3,7 +3,6 @@ package ru.geekbrains.base;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.math.Vector2;
-
 import ru.geekbrains.math.Rect;
 
 public class Sprite extends Rect {
