+++ core/src/ru/geekbrains/screen/GameScreen.java
@@ -1,56 +1,71 @@
 package ru.geekbrains.screen;
 
 import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.math.Vector2;
-import ru.geekbrains.ScreenController;
+import ru.geekbrains.controllers.ScreenController;
 import ru.geekbrains.base.BaseScreen;
 import ru.geekbrains.base.Layer;
 import ru.geekbrains.math.Rect;
-import ru.geekbrains.sprite.Background;
-import ru.geekbrains.sprite.Cloud;
-import ru.geekbrains.sprite.PauseButton;
-import ru.geekbrains.sprite.Player;
-import java.util.concurrent.ThreadLocalRandom;
+import ru.geekbrains.sprite.gameObjects.Background;
+import ru.geekbrains.sprite.gameObjects.Cloud;
+import ru.geekbrains.sprite.buttons.PauseButton;
+import ru.geekbrains.sprite.gameObjects.Player;
 
 public class GameScreen extends BaseScreen {
 
     private Texture bg;
-    private Texture buttonPause;
-    private Texture playerTexture;
+    private TextureAtlas atlas;
     private Background background;
     private Player player;
     private PauseButton pauseButton;
-    private Texture cloudTexture;
+
+    private TextureAtlas.AtlasRegion regionBackground;
+    private TextureAtlas.AtlasRegion regionPlayer;
+    private TextureAtlas.AtlasRegion regionButtonPause;
+    private TextureAtlas.AtlasRegion cloudTextureRegion;
+
+
     private Cloud[] cloudsForeground;
+    private Cloud[] cloudsMiddle;
     private Cloud[] cloudsBackground;
+
     private final int FOREGROUND_CLOUDS_COUNT = 5;
+    private final int MIDDLE_CLOUDS_COUNT = 7;
     private final int BACKGROUND_CLOUDS_COUNT = 10;
 
-    public GameScreen(ScreenController controller) {
+    public GameScreen(TextureAtlas atlas, ScreenController controller) {
         super(controller);
+        this.atlas = atlas;
         bg = new Texture("textures/sky.jpg");
-        playerTexture = new Texture("textures/plane1.png");
-        buttonPause = new Texture("textures/buttonPause.png");
+        regionPlayer = new TextureAtlas.AtlasRegion(atlas.findRegion("plane1"));
+        regionButtonPause = new TextureAtlas.AtlasRegion(atlas.findRegion("buttonPause"));
+
         background = new Background(bg);
-        player = new Player(playerTexture);
-        pauseButton = new PauseButton(buttonPause, controller);
+        player = new Player(regionPlayer);
+        pauseButton = new PauseButton(regionButtonPause, controller);
+
         cloudsForeground = new Cloud[FOREGROUND_CLOUDS_COUNT];
+        cloudsMiddle = new Cloud[MIDDLE_CLOUDS_COUNT];
         cloudsBackground = new Cloud[BACKGROUND_CLOUDS_COUNT];
         for (int i = 0; i < FOREGROUND_CLOUDS_COUNT; i++) {
-            cloudTexture = getRandomCloudTexture();
-            cloudsForeground[i] = new Cloud(cloudTexture, Layer.FOREGROUND);
+            cloudTextureRegion = getRandomCloudTexture();
+            cloudsForeground[i] = new Cloud(cloudTextureRegion, Layer.FOREGROUND);
+        }
+        for (int i = 0; i < MIDDLE_CLOUDS_COUNT; i++) {
+            cloudTextureRegion = getRandomCloudTexture();
+            cloudsMiddle[i] = new Cloud(cloudTextureRegion, Layer.MIDDLE);
         }
         for (int i = 0; i < BACKGROUND_CLOUDS_COUNT; i++) {
-            cloudTexture = getRandomCloudTexture();
-            cloudsBackground[i] = new Cloud(cloudTexture, Layer.BACKGROUND);
+            cloudTextureRegion = getRandomCloudTexture();
+            cloudsBackground[i] = new Cloud(cloudTextureRegion, Layer.BACKGROUND);
         }
 
     }
 
-    private Texture getRandomCloudTexture() {
-
-        String path = "textures/cloud" + (int)((Math.random() * 4) + 1) + ".png";
-        return new Texture(path);
+    private TextureAtlas.AtlasRegion getRandomCloudTexture() {
+        String path = "cloud" + (int)((Math.random() * 4) + 1);
+        return new TextureAtlas.AtlasRegion(atlas.findRegion(path));
     }
 
     @Override
@@ -62,6 +77,9 @@ public class GameScreen extends BaseScreen {
     public void resize(Rect worldBounds) {
         background.resize(worldBounds);
         for (Cloud cloud: cloudsBackground) {
+            cloud.resize(worldBounds);
+        }
+        for (Cloud cloud: cloudsMiddle) {
             cloud.resize(worldBounds);
         }
         player.resize(worldBounds);
@@ -79,6 +97,9 @@ public class GameScreen extends BaseScreen {
         for (Cloud cloud: cloudsBackground) {
             cloud.draw(batch);
         }
+        for (Cloud cloud: cloudsMiddle) {
+            cloud.draw(batch);
+        }
         player.draw(batch);
         for (Cloud cloud: cloudsForeground) {
             cloud.draw(batch);
@@ -89,10 +110,6 @@ public class GameScreen extends BaseScreen {
 
     @Override
     public void dispose() {
-        bg.dispose();
-        buttonPause.dispose();
-        playerTexture.dispose();
-        cloudTexture.dispose();
         super.dispose();
     }
