--- core/src/ru/geekbrains/sprite/Background.java
@@ -1,20 +1,0 @@
-package ru.geekbrains.sprite;
-
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
-
-import ru.geekbrains.base.Sprite;
-import ru.geekbrains.math.Rect;
-
-public class Background extends Sprite {
-
-    public Background(Texture texture) {
-        super(new TextureRegion(texture));
-    }
-
-    @Override
-    public void resize(Rect worldBounds) {
-        setHeightProportion(1.1f);
-        this.pos.set(worldBounds.pos);
-    }
-}
