+++ core/src/ru/geekbrains/sprite/gameObjects/Background.java
@@ -1,0 +1,19 @@
+package ru.geekbrains.sprite.gameObjects;
+
+import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.graphics.g2d.TextureRegion;
+import ru.geekbrains.base.Sprite;
+import ru.geekbrains.math.Rect;
+
+public class Background extends Sprite{
+
+    public Background(Texture texture) {
+        super(new TextureRegion(texture));
+    }
+
+    @Override
+    public void resize(Rect worldBounds) {
+        setHeightProportion(1.1f);
+        this.pos.set(worldBounds.pos);
+    }
+}
